Refactor: Extract control helpers to blinx.controls.js
Move shared control logic into a new module for better organization and reuse.
