+++ lib/blinx.collection.js
@@ -1,5 +1,10 @@
 import { EventTypes } from './blinx.store.js';
 import { RegisteredUI } from './blinx.registered-ui.js';
+import {
+  normalizeControlSpecEntry,
+  resolveElementByIdWithinRoot,
+  applyControlPresentation,
+} from './blinx.controls.js';
 
 const ROOT_CLEANUP = Symbol('blinxCollectionCleanup');
 
@@ -11,60 +16,6 @@ const BUILTIN_CONTROL_KEYS = [
   'nextButton',
   'pageLabel',
 ];
-
-function assertNoLegacyControlIds(obj, source = 'controls') {
-  if (!obj || typeof obj !== 'object') return;
-  const legacyKeys = [
-    'createButtonId',
-    'deleteSelectedButtonId',
-    'statusId',
-    'prevButtonId',
-    'nextButtonId',
-    'pageLabelId',
-  ].filter(k => Object.prototype.hasOwnProperty.call(obj, k));
-  if (legacyKeys.length > 0) {
-    throw new Error(
-      `blinxCollection: legacy control id keys are no longer supported (${legacyKeys.join(', ')}). `
-      + `Use ${source}.{ createButton, deleteSelectedButton, status, prevButton, nextButton, pageLabel } `
-      + 'with shorthand: true | false | "domId" | { domId, label, visible, disabled, css, action }.'
-    );
-  }
-}
-
-function normalizeControlSpecEntry(entry) {
-  if (typeof entry === 'string') return { domId: entry };
-  if (entry === true) return { visible: true, disabled: false };
-  if (entry === false) return { visible: false, disabled: true };
-  if (entry && typeof entry === 'object') return { ...entry };
-  return null;
-}
-
-function resolveElementByIdWithinRoot(root, id) {
-  if (!id) return null;
-  if (root && typeof root.querySelector === 'function') {
-    try {
-      const escaped = (typeof CSS !== 'undefined' && typeof CSS.escape === 'function') ? CSS.escape(id) : id;
-      const within = root.querySelector(`#${escaped}`);
-      if (within) return within;
-    } catch { /* ignore */ }
-  }
-  return document.getElementById(id);
-}
-
-function applyControlPresentation(el, spec) {
-  if (!el || !spec) return;
-  if (typeof spec.visible === 'boolean') el.hidden = !spec.visible;
-  if (typeof spec.disabled === 'boolean') {
-    if ('disabled' in el) el.disabled = !!spec.disabled;
-    el.setAttribute('aria-disabled', spec.disabled ? 'true' : 'false');
-  }
-  if (typeof spec.css === 'string' && spec.css.trim()) {
-    el.className = `${el.className || ''} ${spec.css}`.trim();
-  }
-  if (typeof spec.label === 'string') {
-    if (el.tagName === 'BUTTON') el.textContent = spec.label;
-  }
-}
 
 function defaultBuiltinControlMeta(key) {
   switch (key) {
@@ -378,7 +329,6 @@ export function blinxCollection({
   const explicitControlsProvided = Object.prototype.hasOwnProperty.call(arguments?.[0] || {}, 'controls');
   const viewControlsSpec = view && typeof view === 'object' ? view.controls : undefined;
   const resolvedControlsSpec = explicitControlsProvided ? controls : viewControlsSpec;
-  assertNoLegacyControlIds(resolvedControlsSpec, explicitControlsProvided ? 'controls' : 'view.controls');
 
   const dom = {
     createBtn: null,
@@ -395,7 +345,8 @@ export function blinxCollection({
     if (internalToolbar) return internalToolbar;
     internalToolbar = document.createElement('div');
     internalToolbar.className = 'flex blinx-controls';
-    root.insertBefore(internalToolbar, contentRoot);
+    // Keep placement consistent with auto-rendered toolbar: after content.
+    root.appendChild(internalToolbar);
     cleanupFns.push(() => {
       try { internalToolbar?.remove(); } catch { /* ignore */ }
       internalToolbar = null;
