+++ lib/blinx.form.js
@@ -2,6 +2,11 @@
 import { validateField } from './blinx.validate.js';
 import { EventTypes } from './blinx.store.js';
 import { RegisteredUI } from './blinx.registered-ui.js';
+import {
+  normalizeControlSpecEntry,
+  resolveElementByIdWithinRoot,
+  applyControlPresentation,
+} from './blinx.controls.js';
 
 // Prevent handler accumulation when blinxForm is called multiple times
 // with the same DOM controls (common in tests and re-renders).
@@ -18,74 +23,6 @@ const BUILTIN_CONTROL_KEYS = [
   'recordIndicator',
   'saveStatus',
 ];
-
-function assertNoLegacyControlIds(obj, source = 'controls') {
-  if (!obj || typeof obj !== 'object') return;
-  const legacyKeys = [
-    'saveButtonId',
-    'resetButtonId',
-    'nextButtonId',
-    'prevButtonId',
-    'createButtonId',
-    'deleteButtonId',
-    'recordIndicatorId',
-    'saveStatusId',
-  ].filter(k => Object.prototype.hasOwnProperty.call(obj, k));
-  if (legacyKeys.length > 0) {
-    throw new Error(
-      `blinxForm: legacy control id keys are no longer supported (${legacyKeys.join(', ')}). `
-      + `Use ${source}.{ saveButton, resetButton, nextButton, prevButton, createButton, deleteButton, recordIndicator, saveStatus } `
-      + 'with shorthand: true | false | "domId" | { domId, label, visible, disabled, css, action }.'
-    );
-  }
-}
-
-function normalizeControlSpecEntry(entry) {
-  // Shorthand forms:
-  // - string: domId
-  // - true: visible + auto-render
-  // - false: hidden
-  if (typeof entry === 'string') return { domId: entry };
-  if (entry === true) return { visible: true, disabled: false };
-  if (entry === false) return { visible: false, disabled: true };
-  if (entry && typeof entry === 'object') return { ...entry };
-  // undefined/null => treat as omitted (not declared)
-  return null;
-}
-
-function resolveElementByIdWithinRoot(root, id) {
-  if (!id) return null;
-  // Prefer scoping to root (important for nested renderers); fall back to global.
-  if (root && typeof root.querySelector === 'function') {
-    try {
-      // CSS.escape is not guaranteed in all runtimes; best-effort.
-      const escaped = (typeof CSS !== 'undefined' && typeof CSS.escape === 'function') ? CSS.escape(id) : id;
-      const within = root.querySelector(`#${escaped}`);
-      if (within) return within;
-    } catch { /* ignore */ }
-  }
-  return document.getElementById(id);
-}
-
-function applyControlPresentation(el, spec) {
-  if (!el || !spec) return;
-  if (typeof spec.visible === 'boolean') {
-    // Use hidden for accessibility & easy toggling.
-    el.hidden = !spec.visible;
-  }
-  if (typeof spec.disabled === 'boolean') {
-    // Only meaningful for interactive controls.
-    if ('disabled' in el) el.disabled = !!spec.disabled;
-    el.setAttribute('aria-disabled', spec.disabled ? 'true' : 'false');
-  }
-  if (typeof spec.css === 'string' && spec.css.trim()) {
-    el.className = `${el.className || ''} ${spec.css}`.trim();
-  }
-  if (typeof spec.label === 'string') {
-    // For buttons, update label.
-    if (el.tagName === 'BUTTON') el.textContent = spec.label;
-  }
-}
 
 function defaultBuiltinControlMeta(key) {
   switch (key) {
@@ -250,7 +187,6 @@ export function blinxForm({
   const explicitControlsProvided = Object.prototype.hasOwnProperty.call(arguments?.[0] || {}, 'controls');
   const viewControlsSpec = view && typeof view === 'object' ? view.controls : undefined;
   const resolvedControlsSpec = explicitControlsProvided ? controls : viewControlsSpec;
-  assertNoLegacyControlIds(resolvedControlsSpec, explicitControlsProvided ? 'controls' : 'view.controls');
 
   const dom = {
     saveBtn: null,
